Rename `payment_gateway` package.
