+++ model/src/main/proto/spine_examples/shareaware/payment_gateway/commands.proto
@@ -31,7 +31,7 @@ package spine_examples.shareaware.payment_gateway;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.shareaware.spine.io";
-option java_package = "io.spine.examples.shareaware.paymentgateway.command";
+option java_package = "io.spine.examples.shareaware.payment_gateway.command";
 option java_outer_classname = "CommandsProto";
 option java_multiple_files = true;
