+++ model/src/main/proto/spine_examples/shareaware/payment_gateway/payment_gateway.proto
@@ -31,7 +31,7 @@ package spine_examples.shareaware.payment_gateway;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.shareaware.spine.io";
-option java_package = "io.spine.examples.shareaware.paymentgateway";
+option java_package = "io.spine.examples.shareaware.payment_gateway";
 option java_outer_classname = "PaymentGatewayProto";
 option java_multiple_files = true;
