+++ model/src/main/proto/spine_examples/shareaware/payment_gateway/rejections.proto
@@ -31,7 +31,7 @@ package spine_examples.shareaware.payment_gateway;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.shareaware.spine.io";
-option java_package = "io.spine.examples.shareaware.paymentgateway.rejection";
+option java_package = "io.spine.examples.shareaware.payment_gateway.rejection";
 option java_multiple_files = false;
 
 import "spine_examples/shareaware/identifiers.proto";
