+++ server/src/test/java/io/spine/examples/shareaware/server/wallet/WalletReplenishmentTest.java
@@ -28,8 +28,8 @@ package io.spine.examples.shareaware.server.wallet;
 
 import io.spine.examples.shareaware.ReplenishmentId;
 import io.spine.examples.shareaware.WalletId;
-import io.spine.examples.shareaware.paymentgateway.command.TransferMoneyFromUser;
-import io.spine.examples.shareaware.paymentgateway.rejection.Rejections.MoneyCannotBeTransferredFromUser;
+import io.spine.examples.shareaware.payment_gateway.command.TransferMoneyFromUser;
+import io.spine.examples.shareaware.payment_gateway.rejection.Rejections.MoneyCannotBeTransferredFromUser;
 import io.spine.examples.shareaware.server.TradingContext;
 import io.spine.examples.shareaware.server.given.GivenMoney;
 import io.spine.examples.shareaware.server.paymentgateway.PaymentGatewayProcess;
